fwd.hpp : do not include <pinocchio/config.hpp> header anymore
+ not used to define macros

Remove header eigen-macros.hpp, integrate into math.hpp
